https://jira.aplana.com/browse/SBRFNDFL-3018 Настройки подразделений. Не удаляется настройка, если нет ни одной строки в ней

+++ gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/departmentconfigproperty/server/DeleteConfigPropertyHandler.java
@@ -85,7 +85,9 @@ public class DeleteConfigPropertyHandler extends AbstractActionHandler<DeleteCon
             }
 
             if (period.getCalendarStartDate().equals(recordVersion.getVersionStart())) {
-                providerSlave.deleteRecordVersions(logger, deleteSlaveList, false);
+                if (!deleteSlaveList.isEmpty()) {
+                    providerSlave.deleteRecordVersions(logger, deleteSlaveList, false);
+                }
                 provider.deleteRecordVersions(logger, deleteList, false);
             } else {
                 providerSlave.updateRecordsVersionEnd(logger, addDayToDate(period.getCalendarStartDate(),-2), deleteSlaveList);
